% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_property_names}
\alias{get_property_names}
\title{Extract property names from a GeoJSON file.}
\usage{
get_property_names(geojson_path)
}
\description{
Extract property names from a GeoJSON file.
}
